Move return button to themed components
