--- src/components/returnButton/index.ts
@@ -1,12 +1,0 @@
-import styled from 'styled-components';
-import { LinkButton } from '../linkButton';
-import { MID_GREEN } from '../../utils/colors';
-
-const ReturnButton = styled(LinkButton)`
-  height: 45px;
-  border-color: ${MID_GREEN};
-  font-size: 18px;
-  color: ${MID_GREEN};
-`;
-
-export default ReturnButton;
